// src/app/main/model/newCommittee.ts
export class NewCommittee{
    ID: string;
    Code: string;
    Logo: any;
    CommitteeName: string;
    City: string;
    StartShift1: string;
    StartShift2: string;
    StartShift3: string;
    StartShift1Holiday: string;
    StartShift2Holiday: string;
    StartShift3Holiday: string;
    Email: string;
    AlertActive: boolean;
}